+++ tests/tests/RawCodeCoverageDataTest.php
@@ -78,48 +78,6 @@ final class RawCodeCoverageDataTest extends TestCase
         $this->assertEquals($lineData, $dataObject->lineCoverage());
     }
 
-    /**
-     * In the path-coverage XDebug format for Xdebug < 2.9.6, the line data exists inside a "lines" array key where the
-     * file has classes or functions. For files without them, the data is stored in the line-only format.
-     */
-    public function testLineDataFromMixedCoverageXDebugFormat(): void
-    {
-        $rawDataFromDriver = [
-            '/some/path/SomeClass.php' => [
-                'lines' => [
-                    8  => 1,
-                    9  => -2,
-                    13 => -1,
-                ],
-                'functions' => [
-
-                ],
-            ],
-            '/some/path/justAScript.php' => [
-                18  => 1,
-                19  => -2,
-                113 => -1,
-            ],
-        ];
-
-        $lineData = [
-            '/some/path/SomeClass.php' => [
-                8  => 1,
-                9  => -2,
-                13 => -1,
-            ],
-            '/some/path/justAScript.php' => [
-                18  => 1,
-                19  => -2,
-                113 => -1,
-            ],
-        ];
-
-        $dataObject = RawCodeCoverageData::fromXdebugWithMixedCoverage($rawDataFromDriver);
-
-        $this->assertEquals($lineData, $dataObject->lineCoverage());
-    }
-
     public function testClear(): void
     {
         $lineDataFromDriver = [